Tests completed

+++ pagantis-finance/src/main/resources/db/migration/V1/V1_2__create_account_table.sql
@@ -2,7 +2,7 @@
 CREATE TABLE ACCOUNT (
   id bigint(20) NOT NULL  AUTO_INCREMENT,
   account_number varchar(500) NOT NULL,
-  balance double NOT NULL DEFAULT '0',
+  balance decimal(20,2) NOT NULL DEFAULT '0',
   bank_customer_id bigint(20) NOT NULL,
   PRIMARY KEY (id),
   UNIQUE KEY account_number (account_number),
